+++ src/main/java/com/hallym/festival/domain/comment/controller/CommentController.java
@@ -3,7 +3,7 @@ package com.hallym.festival.domain.comment.controller;
 import com.hallym.festival.domain.comment.dto.CommentPasswordDto;
 import com.hallym.festival.domain.comment.dto.CommentRequestDto;
 import com.hallym.festival.domain.comment.dto.CommentResponseDto;
-import com.hallym.festival.domain.comment.service.CommentService;
+import com.hallym.festival.domain.comment.service.CommentServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
@@ -14,24 +14,24 @@ import java.util.Map;
 @RequestMapping("/comment")
 @RequiredArgsConstructor
 public class CommentController {
-    private final CommentService commentService;
+    private final CommentServiceImpl commentServiceImpl;
 
     @GetMapping("{bno}")
     public List<CommentResponseDto> getCommentList(@PathVariable(name = "bno") Long boothId) throws Exception {
-        return commentService.getAll(boothId);
+        return commentServiceImpl.getAll(boothId);
     }
 
     @PostMapping("{bno}")
     public Map<String, String> createComment
             (@PathVariable(name="bno") Long boothId,
              @RequestBody CommentRequestDto commentRequestDto, HttpServletRequest request){
-        commentService.create(boothId, commentRequestDto, request);
+        commentServiceImpl.create(boothId, commentRequestDto, request);
         return Map.of("result", "create success");
     }
 
     @DeleteMapping("{cno}")
     public Map<String, String> deleteComment(@PathVariable(name = "cno") Long commentId, @RequestBody CommentPasswordDto pwd) {
-        String result = commentService.delete(commentId, pwd);
+        String result = commentServiceImpl.delete(commentId, pwd);
         return Map.of("result", result);
     }
