+++ src/main/java/com/hallym/festival/domain/likes/controller/LikeController.java
@@ -1,30 +1,28 @@
 package com.hallym.festival.domain.likes.controller;
 
-import com.hallym.festival.domain.likes.service.LikeService;
+import com.hallym.festival.domain.likes.service.LikeServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Map;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/like")
 @RequiredArgsConstructor
 public class LikeController {
-    private final LikeService likeService;
+    private final LikeServiceImpl likeServiceImpl;
 
     @GetMapping("{bno}")
     public Map<String, Integer> getLikeCnt(@PathVariable Long bno) {
-        int cnt = likeService.getCount(bno);
+        int cnt = likeServiceImpl.getCount(bno);
         return Map.of("result" , cnt);
     }
 
     @PostMapping("{bno}")
     public Map<String, String> like(@PathVariable Long bno, HttpServletRequest request, HttpServletResponse response) {
-        String result = likeService.likeTrigger(bno, request, response);
+        String result = likeServiceImpl.likeTrigger(bno, request, response);
         return Map.of("result", result);
     }
 }
