+++ src/test/java/com/hallym/festival/service/CommentServiceTest.java
@@ -1,13 +1,9 @@
 package com.hallym.festival.service;
 
-import com.hallym.festival.domain.booth.entity.Booth;
-import com.hallym.festival.domain.booth.entity.BoothType;
-import com.hallym.festival.domain.booth.repository.BoothRepository;
 import com.hallym.festival.domain.comment.dto.CommentPasswordDto;
 import com.hallym.festival.domain.comment.dto.CommentRequestDto;
 import com.hallym.festival.domain.comment.dto.CommentResponseDto;
-import com.hallym.festival.domain.comment.service.CommentService;
-import lombok.RequiredArgsConstructor;
+import com.hallym.festival.domain.comment.service.CommentServiceImpl;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,17 +11,14 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.mock.web.MockHttpServletRequest;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.IntStream;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Log4j2
 class CommentServiceTest {
 
     @Autowired
-    CommentService commentService;
+    CommentServiceImpl commentServiceImpl;
 
     @Test
     public void create() {
@@ -38,7 +31,7 @@ class CommentServiceTest {
                             .password("jyp1234")
                             .content("댓글입니다."+i).build();
 
-            commentService.create(boothId, commentRequestDto, request);
+            commentServiceImpl.create(boothId, commentRequestDto, request);
         });
     }
 
@@ -48,7 +41,7 @@ class CommentServiceTest {
             if(i%2 == 0) {
                 Long boothId = (long) i;
                 CommentPasswordDto pDto = CommentPasswordDto.builder().password("jyp1234").build();
-                commentService.delete(boothId, pDto);
+                commentServiceImpl.delete(boothId, pDto);
             }
         });
     }
@@ -56,7 +49,7 @@ class CommentServiceTest {
     @Test
     public void getAll() throws Exception {
         Long boothId = 1L;
-        List<CommentResponseDto> commentList = commentService.getAll(boothId);
+        List<CommentResponseDto> commentList = commentServiceImpl.getAll(boothId);
         log.info(commentList.size());
 
         for (CommentResponseDto comment : commentList) {
